Extract command base class for loading a schema from inputs
