+++ redwood-tooling-schema/src/main/kotlin/app/cash/redwood/tooling/schema/JsonCommand.kt
@@ -15,62 +15,22 @@
  */
 package app.cash.redwood.tooling.schema
 
-import com.github.ajalt.clikt.core.CliktCommand
-import com.github.ajalt.clikt.parameters.arguments.argument
-import com.github.ajalt.clikt.parameters.arguments.help
-import com.github.ajalt.clikt.parameters.options.defaultLazy
-import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.help
-import com.github.ajalt.clikt.parameters.options.multiple
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
-import com.github.ajalt.clikt.parameters.options.split
-import com.github.ajalt.clikt.parameters.types.file
 import com.github.ajalt.clikt.parameters.types.path
-import java.io.File
-import java.net.URLClassLoader
 import kotlin.io.path.createDirectories
 import kotlin.io.path.writeText
 
 internal class JsonCommand :
-  CliktCommand(
+  AbstractSchemaCommand(
     name = "json",
     help = "Parse schema members into a JSON representation",
   ) {
   private val out by option().path().required()
     .help("Directory into which JSON is written")
 
-  private val fir by option("--fir")
-    .flag()
-    .help("Use new FIR-based parser")
-
-  private val jdkHome by option("--jdk-home")
-    .file()
-    .defaultLazy { System.getProperty("java.home").let(::File) }
-    .help("Path to the JDK installation (defaults to this JDK)")
-
-  private val sources by option("-s", "--source")
-    .file()
-    .multiple()
-    .help("Kotlin source files and folders")
-
-  private val classpath by option("-cp", "--class-path")
-    .file()
-    .split(File.pathSeparator)
-    .required()
-
-  private val schemaTypeName by argument("schema")
-    .help("Fully-qualified class name for the @Schema-annotated interface")
-
   override fun run() {
-    val schema = if (fir) {
-      parseProtocolSchema(jdkHome, sources, classpath, FqType.bestGuess(schemaTypeName)).schema
-    } else {
-      val classLoader = URLClassLoader(classpath.map { it.toURI().toURL() }.toTypedArray())
-      val schemaType = classLoader.loadClass(schemaTypeName).kotlin
-      ProtocolSchemaSet.parse(schemaType).schema
-    }
-
     val embeddedSchema = schema.toEmbeddedSchema()
     val path = out.resolve(embeddedSchema.path)
     path.parent.createDirectories()
